[+] b034362e - updates to menu and clean up of code
