+++ GameFiles/Source/PlayerController.lua
@@ -10,7 +10,7 @@ function Player_Load()
   
   circle = {
     x = 250,
-    y = 800,
+    y = 910,
     radiusx = 30,
     radiusy = 30 }
   
@@ -29,7 +29,6 @@ function Player_Update(dt)
   -- Set circle position to where click occured
   if dragging == true then
     circle.x = love.mouse.getX()
-    circle.y = love.mouse.getY()
   end
   
   if circleRectangleCollide(circle.x, circle.y, circle.radiusx, rectangle.x, rectangle.y, rectangle.width, rectangle.height) then
@@ -51,27 +50,16 @@ function Player_Draw()
   end
 end
 
-
-function love.mousepressed(x, y, button)
-  -- Check if left mouse button pressed, if it is, set dragging to true
-  if gamestate == "game" then
-    if mouseRectCollide(x, y, button, rect.x, rect.y, rect.width, rect.height, 1) then
-      dragging = true
-    end
-    
-  elseif gamestate == "startScreen" then
-    if mouseRectCollide(x, y, button, 70, 200, 400, 130, 1) then
-      gamestate = "game"
-    elseif mouseRectCollide(x, y, button, 70, 400, 400, 130, 1) then
-      gamestate = "scores"
-    end
+function Player_mousereleased(x, y, button)
+  -- resets dragging once mouse released
+  if button == 1 then 
+    dragging = false
   end
 end
 
-
-function love.mousereleased(x, y, button)
-  -- resets dragging once mouse released
-  if button == 1 and gamestate == "game" then 
-    dragging = false
+function Player_mousepressed(x, y, button)
+  -- Check if left mouse button pressed, if it is, set dragging to true
+  if mouseRectCollide(x, y, button, rect.x, rect.y, rect.width, rect.height, 1) then
+    dragging = true
   end
 end
